+++ src/EditorFeatures/Core/Diagnostics/AbstractDiagnosticsTaggerProvider.cs
@@ -80,7 +80,7 @@ internal abstract partial class AbstractDiagnosticsTaggerProvider<TTag> : ITagge
 
     public ITagger<T>? CreateTagger<T>(ITextBuffer buffer) where T : ITag
     {
-        using var taggers = TemporaryArray<RoslynTagger<TTag>>.Empty;
+        using var taggers = TemporaryArray<EfficientTagger<TTag>>.Empty;
         foreach (var taggerProvider in _diagnosticsTaggerProviders)
         {
             var innerTagger = taggerProvider.CreateTagger<TTag>(buffer);
